+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `billingAddress` to `ApplePayCard`
 * Create `AndroidPayCardAddressRequest`, `AndroidPayCardOptionsRequest`, and `AndroidPayCardRequest`, and add `AndroidPayCardRequest` to `CustomerRequest`
 * Add `billingAddress` to `AndroidPayCard`
+* Add `merchantAdviceCode` and `merchantAdviceCodeText` fields to `Transaction`
 
 ## 3.22.0
 * Add subscriptions to `SepaDirectDebitAccount`